Agentops/requirements.txt
crewai>=0.28.0
agentops>=0.1.0
groq>=0.4.0
langchain>=0.0.335
langchain-groq>=0.0.1
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
